Remove usePolygonEndowWithdraw
